requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.2.0
crypto==1.4.1
cryptography==39.0.0
cytoolz==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.10.3
frozenlist==1.3.3
fuzzywuzzy==0.18.0
h11==0.14.0
hexbytes==0.3.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
ipykernel==6.26.0
ipython==8.18.0
jedi==0.19.1
jsonschema==4.17.3
jupyter_client==8.6.0
jupyter_core==5.5.0
lru-dict==1.1.8
matplotlib-inline==0.1.6
multidict==6.0.4
mysql-connector==2.2.9
Naked==0.1.32
nest-asyncio==1.5.8
openai==1.3.5
packaging==23.2
parsimonious==0.9.0
parso==0.8.3
platformdirs==3.1.1
prompt-toolkit==3.0.41
protobuf==4.22.1
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.17
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==305
PyYAML==6.0
pyzmq==25.1.1
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tiktoken==0.5.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.12
web3==6.0.0
websockets==10.4
yarl==1.8.2
